refactor: use `RelativeTime` instead of `TimeAgo`
