+++ packages/sanity/src/core/form/inputs/ReferenceInput/ReferencePreview.tsx
@@ -2,12 +2,12 @@ import React, {useMemo} from 'react'
 import {ObjectSchemaType} from '@sanity/types'
 import {Box, Flex, Inline, Label, Text, Tooltip, useRootTheme} from '@sanity/ui'
 import {EditIcon, PublishIcon} from '@sanity/icons'
+import {RelativeTime} from '../../../components/RelativeTime'
 import {Translate, useTranslation} from '../../../i18n'
 import {RenderPreviewCallback} from '../../types'
 import {PreviewLayoutKey, TextWithTone} from '../../../components'
 import {useDocumentPresence} from '../../../store'
 import {DocumentPreviewPresence} from '../../../presence'
-import {TimeAgo} from './utils/TimeAgo'
 import {ReferenceInfo} from './types'
 
 /**
@@ -85,7 +85,11 @@ export function ReferencePreview(props: {
                         <Translate
                           t={t}
                           i18nKey="inputs.reference.preview.published-at-time"
-                          components={{TimeAgo: () => <TimeAgo time={publishedAt} />}}
+                          components={{
+                            RelativeTime: () => (
+                              <RelativeTime time={publishedAt} useTemporalPhrase />
+                            ),
+                          }}
                         />
                       ) : (
                         <>{t('inputs.reference.preview.not-published')}</>
@@ -121,7 +125,11 @@ export function ReferencePreview(props: {
                         <Translate
                           t={t}
                           i18nKey="inputs.reference.preview.edited-at-time"
-                          components={{TimeAgo: () => <TimeAgo time={draftEditedAt} />}}
+                          components={{
+                            RelativeTime: () => (
+                              <RelativeTime time={draftEditedAt} useTemporalPhrase />
+                            ),
+                          }}
                         />
                       ) : (
                         <>{t('inputs.reference.preview.no-unpublished-edits')}</>
