+++ packages/sanity/src/core/hooks/__tests__/useRelativeTime.test.tsx
@@ -1,12 +1,12 @@
 import React, {type ReactElement} from 'react'
 import {ThemeProvider, studioTheme} from '@sanity/ui'
 import {renderHook} from '@testing-library/react'
-import {useTimeAgo} from '../useTimeAgo'
+import {useRelativeTime} from '../useRelativeTime'
 import {LocaleProviderBase, usEnglishLocale} from '../../i18n'
 import {prepareI18n} from '../../i18n/i18nConfig'
 import {studioDefaultLocaleResources} from '../../i18n/bundles/studio'
 
-describe('useTimeAgo', () => {
+describe('useRelativeTime', () => {
   const timeZone = 'UTC'
   const {i18next} = prepareI18n({
     projectId: 'test',
@@ -36,7 +36,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "Jun 3" if within same year, but different month', () => {
       const date = new Date('2023-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('Jun 3')
@@ -45,7 +45,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "Jun 3, 2022" if different year', () => {
       const date = new Date('2022-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('Jun 3, 2022')
@@ -54,7 +54,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "3 wks" if less than a month apart', () => {
       const date = new Date('2023-07-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('3 wks')
@@ -63,7 +63,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "5 days" if less than a week apart', () => {
       const date = new Date('2023-07-20T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('5 days')
@@ -72,7 +72,7 @@ describe('useTimeAgo', () => {
     it('outputs "yesterday" if one day apart, and in the past', () => {
       const date = new Date('2023-07-24T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('yesterday')
@@ -81,7 +81,7 @@ describe('useTimeAgo', () => {
     it('outputs "tomorrow" if one day apart, and in the future', () => {
       const relativeTo = new Date('2023-07-24T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('tomorrow')
@@ -90,7 +90,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "5 hr" if hours apart', () => {
       const date = new Date('2023-07-25T07:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('5 hr')
@@ -99,7 +99,7 @@ describe('useTimeAgo', () => {
     it('outputs eg "5 min" if minutes apart', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:05:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('5 min')
@@ -108,14 +108,14 @@ describe('useTimeAgo', () => {
     it('outputs eg "20 sec" if seconds apart (and more than 10)', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:20Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, minimal}), {
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo, minimal}), {
         wrapper,
       })
       expect(result.current).toBe('20 sec')
     })
 
     it('outputs "just now" if less than 10 seconds ago', () => {
-      const {result} = renderHook(() => useTimeAgo(new Date()), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(new Date()), {wrapper})
       expect(result.current).toBe('just now')
     })
   })
@@ -127,7 +127,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('Jun 3')
@@ -137,7 +137,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2022-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('Jun 3, 2022')
@@ -147,7 +147,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('3 wk. ago')
@@ -157,7 +157,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-03T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('in 3 wk.')
@@ -167,7 +167,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-20T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('5 days ago')
@@ -177,7 +177,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-20T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('in 5 days')
@@ -187,7 +187,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-24T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('yesterday')
@@ -197,7 +197,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-24T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('tomorrow')
@@ -207,7 +207,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-25T07:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('5 hr. ago')
@@ -217,7 +217,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-25T07:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('in 5 hr.')
@@ -227,7 +227,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:05:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('5 min. ago')
@@ -237,7 +237,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:05:00Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('in 5 min.')
@@ -247,7 +247,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:20Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('20 sec. ago')
@@ -257,7 +257,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:00:20Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('in 20 sec.')
@@ -267,7 +267,7 @@ describe('useTimeAgo', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:05Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('just now')
@@ -277,7 +277,7 @@ describe('useTimeAgo', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:00:05Z')
       const {result} = renderHook(
-        () => useTimeAgo(date, {timeZone, relativeTo, minimal, agoSuffix: true}),
+        () => useRelativeTime(date, {timeZone, relativeTo, minimal, useTemporalPhrase: true}),
         {wrapper},
       )
       expect(result.current).toBe('just now')
@@ -288,82 +288,82 @@ describe('useTimeAgo', () => {
     it('outputs full date, eg "Jun 3, 2023, 2:00 PM" if within same year, but different month', () => {
       const date = new Date('2023-06-03T14:00:00Z')
       const relativeTo = new Date('2023-07-25T14:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('Jun 3, 2023, 2:00 PM')
     })
 
     it('outputs full date eg "Jun 3, 2022, 2:00 PM" if different year', () => {
       const date = new Date('2022-06-03T14:00:00Z')
       const relativeTo = new Date('2023-07-25T14:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('Jun 3, 2022, 2:00 PM')
     })
 
     it('outputs eg "3 weeks" if less than a month apart', () => {
       const date = new Date('2023-07-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('3 weeks')
     })
 
     it('outputs eg "1 week" if less than a month apart', () => {
       const date = new Date('2023-07-17T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('1 week')
     })
 
     it('outputs eg "5 days" if less than a week apart', () => {
       const date = new Date('2023-07-20T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('5 days')
     })
 
     it('outputs "yesterday" if one day apart', () => {
       const date = new Date('2023-07-24T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('yesterday')
     })
 
     it('outputs eg "5 hours" if hours apart', () => {
       const date = new Date('2023-07-25T07:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('5 hours')
     })
 
     it('outputs "1 hour" if an hour apart', () => {
       const date = new Date('2023-07-25T11:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('1 hour')
     })
 
     it('outputs eg "5 minutes" if minutes apart', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:05:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('5 minutes')
     })
 
     it('outputs "1 minute" if a minute apart', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:01:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('1 minute')
     })
 
     it('outputs eg "20 seconds" if seconds apart (and more than 10)', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:20Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo}), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(date, {timeZone, relativeTo}), {wrapper})
       expect(result.current).toBe('20 seconds')
     })
 
     it('outputs "just now" if less than 10 seconds ago', () => {
-      const {result} = renderHook(() => useTimeAgo(new Date()), {wrapper})
+      const {result} = renderHook(() => useRelativeTime(new Date()), {wrapper})
       expect(result.current).toBe('just now')
     })
   })
@@ -372,198 +372,264 @@ describe('useTimeAgo', () => {
     it('outputs full date, eg "Jun 3, 2023 12:00 PM" if within same year, but different month', () => {
       const date = new Date('2023-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('Jun 3, 2023, 12:00 PM')
     })
 
     it('outputs full date, eg "Jun 3, 2022 12:00 PM" if different year', () => {
       const date = new Date('2022-06-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('Jun 3, 2022, 12:00 PM')
     })
 
     it('outputs eg "3 weeks ago" if less than a month apart, and in the past', () => {
       const date = new Date('2023-07-03T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('3 weeks ago')
     })
 
     it('outputs eg "in 3 weeks" if less than a month apart, and in the future', () => {
       const relativeTo = new Date('2023-07-03T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 3 weeks')
     })
 
     it('outputs eg "last week" if diff is a week, and in the past', () => {
       const date = new Date('2023-07-17T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('last week')
     })
 
     it('outputs eg "next week" if diff is a week, and in the future', () => {
       const relativeTo = new Date('2023-07-17T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('next week')
     })
 
     it('outputs eg "5 days ago" if less than a week apart, and in the past', () => {
       const date = new Date('2023-07-20T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('5 days ago')
     })
 
     it('outputs eg "in 5 days" if less than a week apart, and in the future', () => {
       const relativeTo = new Date('2023-07-20T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 5 days')
     })
 
     it('outputs "yesterday" if one day apart, and in the past', () => {
       const date = new Date('2023-07-24T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('yesterday')
     })
 
     it('outputs "tomorrow" if one day apart, and in the future', () => {
       const relativeTo = new Date('2023-07-24T12:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('tomorrow')
     })
 
     it('outputs eg "5 hours ago" if hours apart, and in the past', () => {
       const date = new Date('2023-07-25T07:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('5 hours ago')
     })
 
     it('outputs eg "in 5 hours" if hours apart, and in the future', () => {
       const relativeTo = new Date('2023-07-25T07:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 5 hours')
     })
 
     it('outputs eg "1 hours ago" if an hour apart, and in the past', () => {
       const date = new Date('2023-07-25T11:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('1 hour ago')
     })
 
     it('outputs eg "in 1 hour" if an hour apart, and in the future', () => {
       const relativeTo = new Date('2023-07-25T11:00:00Z')
       const date = new Date('2023-07-25T12:00:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 1 hour')
     })
 
     it('outputs eg "5 minutes ago" if minutes apart, and in the past', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:05:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('5 minutes ago')
     })
 
     it('outputs eg "in 5 minutes" if minutes apart, and in the future', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:05:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 5 minutes')
     })
 
     it('outputs eg "1 minute ago" if a minute apart, and in the past', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:01:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('1 minute ago')
     })
 
     it('outputs eg "in 1 minute" if a minute apart, and in the future', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:01:00Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 1 minute')
     })
 
     it('outputs eg "20 seconds ago" if more than 10 seconds apart, and in the past', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:20Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('20 seconds ago')
     })
 
     it('outputs eg "in 20 seconds" if more than 10 seconds apart, and in the future', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:00:20Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('in 20 seconds')
     })
 
     it('outputs "just now" if less than 10 seconds ago', () => {
       const date = new Date('2023-07-25T12:00:00Z')
       const relativeTo = new Date('2023-07-25T12:00:05Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('just now')
     })
 
     it('outputs "just now" if less than 10 seconds in the future', () => {
       const relativeTo = new Date('2023-07-25T12:00:00Z')
       const date = new Date('2023-07-25T12:00:05Z')
-      const {result} = renderHook(() => useTimeAgo(date, {timeZone, relativeTo, agoSuffix: true}), {
-        wrapper,
-      })
+      const {result} = renderHook(
+        () => useRelativeTime(date, {timeZone, relativeTo, useTemporalPhrase: true}),
+        {
+          wrapper,
+        },
+      )
       expect(result.current).toBe('just now')
     })
   })
