+++ packages/sanity/src/core/hooks/index.ts
@@ -5,6 +5,7 @@ export * from './useDocumentOperation'
 export * from './useDocumentOperationEvent'
 export * from './useEditState'
 export * from './useProjectId'
+export * from './useRelativeTime'
 export * from './useSchema'
 export * from './useSyncState'
 export * from './useTemplates'
