+++ packages/sanity/src/desk/components/TimeAgo.tsx
@@ -1,12 +1,16 @@
 import React from 'react'
-import {useTimeAgo} from 'sanity'
+import {useRelativeTime} from 'sanity'
 
 export interface TimeAgoProps {
   time: string | Date
 }
 
+/**
+ * @deprecated Use {@link RelativeTime} instead
+ * @internal
+ */
 export function TimeAgo({time}: TimeAgoProps) {
-  const timeAgo = useTimeAgo(time)
+  const timeAgo = useRelativeTime(time)
 
   return <span title={timeAgo}>{timeAgo} ago</span>
 }
